fix: extract table details to component
because the reusable template didn't have any props
it was not properly reactive.
